refactor: relocate NotificationCenter.tsx from Navbar/ to notifications/
