+++ apps/web/src/components/layout/Navbar/NavbarMobileDrawer.tsx
@@ -48,7 +48,7 @@ import { Button } from '@/components/ui/primitives/button';
 import { useCurrentUser } from '@/hooks/queries/useCurrentUser';
 import { cn } from '@/lib/utils';
 
-import { Logo } from './Logo';
+import { Logo } from '@/components/layout/TopNavbar/Logo';
 
 // ─── Nav Section Items ───────────────────────────────────────────────────────
