+++ apps/web/src/components/layout/Navbar/index.ts
@@ -8,15 +8,27 @@
 
 // ─── New 3-section Navbar (Issue #5036) ──────────────────────────────────────
 export { Navbar, type NavbarProps } from './Navbar';
-export { NavbarDropdown, type NavbarDropdownProps, type NavbarDropdownItem } from './NavbarDropdown';
+export {
+  NavbarDropdown,
+  type NavbarDropdownProps,
+  type NavbarDropdownItem,
+} from './NavbarDropdown';
 export { NavbarMobileDrawer, type NavbarMobileDrawerProps } from './NavbarMobileDrawer';
 export { NavbarUserMenu, type NavbarUserMenuProps } from './NavbarUserMenu';
 
 // ─── Legacy sub-components (kept for migration period) ────────────────────────
 export { HamburgerButton, type HamburgerButtonProps } from './HamburgerButton';
 export { HamburgerMenu, type HamburgerMenuProps } from './HamburgerMenu';
-export { Logo, type LogoProps } from './Logo';
-export { NavItems, NavItemButton, type NavItem, type NavItemsProps, type NavItemButtonProps } from './NavItems';
+export {
+  NavItems,
+  NavItemButton,
+  type NavItem,
+  type NavItemsProps,
+  type NavItemButtonProps,
+} from './NavItems';
 export { ProfileBar, type ProfileBarProps } from './ProfileBar';
 export { UniversalNavbar, type UniversalNavbarProps } from './UniversalNavbar';
-export { NotificationCenter, type NotificationCenterProps } from './NotificationCenter';
+export {
+  NotificationCenter,
+  type NotificationCenterProps,
+} from '@/components/notifications/NotificationCenter';
