+++ apps/web/src/components/layout/TopNavbar/index.ts
@@ -1,2 +1,4 @@
 export { TopNavbar } from './TopNavbar';
 export type { TopNavbarProps } from './TopNavbar';
+export { Logo } from './Logo';
+export type { LogoProps } from './Logo';
