+++ apps/web/src/components/notifications/__tests__/NotificationBell.test.tsx
@@ -18,7 +18,7 @@ import { NotificationBell } from '../NotificationBell';
 import { useNotificationStore } from '@/store/notification/store';
 
 // Mock NotificationCenter to avoid IntlProvider requirement in bell unit tests
-vi.mock('@/components/layout/Navbar/NotificationCenter', () => ({
+vi.mock('@/components/notifications/NotificationCenter', () => ({
   NotificationCenter: () => null,
 }));
 
@@ -29,9 +29,11 @@ vi.mock('@/hooks/useNotificationSSE', () => ({
 }));
 
 // Mock NotificationCenter Sheet to isolate bell tests
-vi.mock('@/components/layout/Navbar/NotificationCenter', () => ({
+vi.mock('@/components/notifications/NotificationCenter', () => ({
   NotificationCenter: ({ open }: { open: boolean; onOpenChange: (v: boolean) => void }) =>
-    open ? <div data-testid="notification-center-mock" role="dialog" aria-label="Notifications" /> : null,
+    open ? (
+      <div data-testid="notification-center-mock" role="dialog" aria-label="Notifications" />
+    ) : null,
 }));
 
 // Mock store with all selectors
