+++ apps/web/src/components/notifications/index.ts
@@ -7,4 +7,4 @@
 export { NotificationBell } from './NotificationBell';
 export { NotificationPanel } from './NotificationPanel';
 export { NotificationItem } from './NotificationItem';
-export { NotificationCenter } from '@/components/layout/Navbar/NotificationCenter';
+export { NotificationCenter } from './NotificationCenter';
